chore(dal,sdf): refactor tasks into new module
- Refactor existing tasks (ResourceScheduler and StatusReceiver) into a
  new module in the dal
- Ensure each tasks' respective modules are private (enables us to
  detach this into a separate crate in the future once/if dal cyclic
  dependencies become easier to avoid)
- Add doc comment for "resource_scheduler" module
