+++ lib/dal/src/lib.rs
@@ -47,7 +47,6 @@ pub mod prototype_context;
 pub mod prototype_list_for_func;
 pub mod provider;
 pub mod qualification;
-pub mod resource_scheduler;
 pub mod schema;
 pub mod secret;
 pub mod socket;
@@ -55,7 +54,7 @@ pub mod standard_accessors;
 pub mod standard_model;
 pub mod standard_pk;
 pub mod status;
-pub mod status_receiver;
+pub mod tasks;
 pub mod tenancy;
 pub mod timestamp;
 pub mod user;
@@ -138,7 +137,6 @@ pub use prototype_list_for_func::{
 pub use provider::external::{ExternalProvider, ExternalProviderError, ExternalProviderId};
 pub use provider::internal::{InternalProvider, InternalProviderError, InternalProviderId};
 pub use qualification::{QualificationError, QualificationView};
-pub use resource_scheduler::{ResourceScheduler, ResourceSchedulerError};
 pub use schema::variant::leaves::LeafInput;
 pub use schema::variant::leaves::LeafInputLocation;
 pub use schema::variant::leaves::LeafKind;
